requirements.txt
# Core AI and ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
safetensors>=0.3.1
scikit-learn>=1.3.0
scipy>=1.11.0 # For KDTree in color quantization

# Image processing and data handling
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
omegaconf>=2.3.0
matplotlib>=3.7.0

# Vectorization and GUI support
cairosvg>=2.7.0 # For SVG to PDF conversion

# AI Models from Git repositories
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
controlnet-aux>=0.0.6